Initial commit
This is a basic isotropic hardening, von Mises yield criterion plasticity model in Matlab. The code is roughly taken from the book by Dunne and Petrinic. The Main script, UMAT_Tester.m, reads in a file of total strain increments, dstran and works with this. It is only roughly a uniaxial stress model.